move gql files to new directory
